+++ src/ewatercycle_HBV/model.py
@@ -11,6 +11,23 @@ from ewatercycle_HBV.forcing import HBVForcing # Use custom forcing instead
 from ewatercycle.base.model import ContainerizedModel, eWaterCycleModel
 from ewatercycle.container import ContainerImage
 
+HBV_PARAMS = (
+    "Imax",
+    "Ce",
+    "Sumax",
+    "Beta",
+    "Pmax",
+    "Tlag",
+    "Kf",
+    "Ks",
+)
+
+HBV_STATES = (
+    "Si",
+    "Su",
+    "Sf",
+    "Ss"
+)
 
 class HBVMethods(eWaterCycleModel):
     """
@@ -72,7 +89,50 @@ class HBVMethods(eWaterCycleModel):
 
     @property
     def parameters(self) -> ItemsView[str, Any]:
-        return self._config.items()
+        """List the (initial!) parameters for this model.
+
+        Exposed HBV parameters:
+            Imax (mm): is the maximum amount of interception, under the assumption that all interception evaporates
+
+            Ce(−): is a parameter used to describe what factor actually evaporates from the ground
+                  (Unsaturated root zone: 𝑆𝑢) and is 𝐸𝐴=𝑆𝑢𝑆𝑢,𝑚𝑎𝑥×𝐶𝑒𝐸𝑝
+
+            Sumax(mm): Is the size of the reservoir of the unsaturated root zone (𝑆𝑢)
+                    i.e. the amount of water the top layer of soil can hold. This parameter is used in a few other calculations.
+
+            Beta (−): is a factor controlling the split between fast and slow flow (overland vs groundwater).
+                    Some water will be held by the soil whilst some flows over it and straight to the river.
+                    𝐶𝑟=(𝑆𝑢/𝑆𝑢_𝑚𝑎𝑥)^𝛽 which is used to determine the water infiltrating: 𝑄𝑖𝑢=(1−𝐶𝑟)𝑃𝑒
+                    where 𝑃𝑒 is the actual precipitation reaching the ground.
+                    The rest flow into the fast reservoir 𝑄𝑢𝑓 – which is the groundwater.
+
+            Pmax (mm): is the maximum amount of percolation which can occur from the ground
+                       to the deeper groundwater flow: 𝑄𝑢𝑠=𝑃𝑚𝑎𝑥(𝑆𝑢𝑆𝑢,𝑚𝑎𝑥)
+
+            T_lag (d): is the lag time between water falling and reaching the river.
+
+            Kf (-): the fast flow is modelled as a linear reservoir thus a fraction
+                    of the volume stored leaves to the river 𝑄𝑓=𝐾𝑓∗𝑆𝑓
+
+            Ks (−): Similarly the slow flow is also modelled as 𝑄𝑆=𝐾𝑠∗𝑆𝑆.
+
+        """
+        pars: dict[str, Any] = dict(zip(HBV_PARAMS, self._config["parameters"].split(',')))
+        return pars.items()
+
+    @property
+    def states(self) -> ItemsView[str, Any]:
+        """List the (initial!) states for this model.
+
+        Exposed HBV states:
+            Si (mm): Interception storage, water stored in leaves
+            Su (mm): Unsaturated rootzone storage, water stored accessible to plants
+            Sf (mm): Fastflow storage, moving Fast through the soil - preferential flow paths, upper level
+            Ss (mm): Groundwater storage, moving Slowly through the soil - deeper grounds water.
+
+        """
+        pars: dict[str, Any] = dict(zip(HBV_STATES, self._config["initial_storage"].split(',')))
+        return pars.items()
 
 
     def finalize(self) -> None:
@@ -113,5 +173,5 @@ class HBVMethods(eWaterCycleModel):
 class HBV(ContainerizedModel, HBVMethods):
     """The HBV eWaterCycle model, with the Container Registry docker image."""
     bmi_image: ContainerImage = ContainerImage(
-        "ghcr.io/daafip/hbv-bmi-grpc4bmi:v1.2.0"
+        "ghcr.io/daafip/hbv-bmi-grpc4bmi:v1.3.0"
     )