+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Adding `.finalize()` method - clears up the directory. Especially useful for DA.
 ### v1.4.0
 - Refactor `forcing_file` attribute to `camel_file` -> as essentially it's now a txt file of specific class which the from_camels functions is run on in model startup
 - change selection of dataset period to be inclusive
-### v.1.4.1 
+#### v.1.4.1 
 - changes to config file: alpha is not needed in model, only in forcing per definition. 
 - now removes any netcdf files created when using camels or test data
+#### v.1.4.2
+- updated parameter function to return nicely formated parameter dict
+### 1.5.0
+- changes output model parameter from `Q_m` to `Q`!!! **this is not backward compatible*